# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lewis Bookstore API",
    "version": "1.0.0",
    "description": "RESTful API for managing books, authors, and users."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Retrieve all books",
        "responses": {
          "200": {
            "description": "List of all books"
          }
        }
      },
      "post": {
        "summary": "Add a new book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "authorName": { "type": "string" },
                  "isbn": { "type": "string" },
                  "pubDate": { "type": "string" }
                },
                "required": ["title", "authorName", "isbn", "pubDate"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Book created successfully" },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Get a book by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Book details" },
          "404": { "description": "Book not found" }
        }
      },
      "put": {
        "summary": "Update a book by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "authorName": { "type": "string" },
                  "isbn": { "type": "string" },
                  "pubDate": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Book updated successfully" },
          "400": { "description": "Invalid input" },
          "404": { "description": "Book not found" }
        }
      },
      "delete": {
        "summary": "Delete a book by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Book deleted successfully" },
          "404": { "description": "Book not found" }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Retrieve all authors",
        "responses": {
          "200": { "description": "List of authors" }
        }
      },
      "post": {
        "summary": "Add a new author",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "bio": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Author created successfully" },
          "400": { "description": "Invalid input" }
        }
      }
    }
  }
}